requirements.txt
anthropic
fastapi
uvicorn
chromadb
pydantic
python-dotenv